+++ examples/dequantize_gemm/example_dequant_gemm_bf16_mxfp4_hopper.py
@@ -4,7 +4,7 @@ from tilelang import tvm as tvm
 from tvm import DataType
 from tvm import tir
 import torch
-from utils import torch_convert_bit_twiddling, torch_convert
+from dequantize_utils import torch_convert_bit_twiddling, torch_convert
 
 
 def _tir_u8_to_f4_to_bf16(nbit: int, val: tir.PrimExpr, pos: tir.PrimExpr, scale: tir.PrimExpr,
